Create Digital Clock GUIi
This is a simple Clock GUI application built using Java Swing that displays the current time and date with the option to toggle between 12-hour (AM/PM) and 24-hour formats. The clock updates every second and the user can interact with the two buttons to change the time format dynamically.